api/requirements.txt
Flask~=2.3
psycopg2-binary
Flask-Cors
